requirements.txt
python-telegram-bot>=22.0
yt-dlp>=2025.5.0
ffmpeg-python>=0.2.0
requests>=2.32.0
fastapi>=0.104.0
uvicorn>=0.23.0
